services/detector-awc135/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
timm==1.0.9
pillow==10.4.0
httpx==0.27.0
requests==2.32.3
